requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
kombu==5.4.2
mysqlclient==2.2.6
outcome==1.3.0.post0
packaging==24.2
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
